Improve wrapper to better handle stack projects
The previous wrapper code performed some basic detection of stack
projects, but did not handle hie installed using `stack build
--copy-compiler-tool` unless atom was started using `stack exec --
atom`. This commit adjusts the wrapper script to launch hie using `stack
exec` when a stack project is detected, with the added benefit of
ensuring all executables on the project’s path are on the PATH when hie
runs.

Additionally, this commit makes some minor changes to the way custom
wrappers are handled, treating them closer to the way the default
wrapper is handled. Primarily, this means passing the actual arguments
that should be passed to HIE instead of passing a single empty string
argument (which was not terribly useful).